refactor(esp32): rename bootloader_trigger to dev_dfu for clearer naming
- Rename bootloader_trigger.c/.h to dev_dfu.c/.h
- Update include references in main.c and usb_cdc.c
- Update CMakeLists.txt to reference dev_dfu.c
- Maintains consistent naming convention for device firmware update functionality
