+++ esp32/main/dev_dfu.c
@@ -1,4 +1,4 @@
-#include "bootloader_trigger.h"
+#include "dev_dfu.h"
 
 #include "esp_log.h"
 #include "esp_system.h"
